% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{bic}
\alias{bic}
\title{BIC}
\usage{
bic(e, x, parameters, outlier = FALSE)
}
\arguments{
\item{e}{a function}

\item{x}{a matrix}

\item{parameters}{a list}

\item{outlier}{a boolean}
}
\value{
a number
}
\description{
BIC
}
