% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{e_step}
\alias{e_step}
\title{Expectation step}
\usage{
e_step(x, parameters, outlier = FALSE)
}
\arguments{
\item{x}{a matrix}

\item{parameters}{a list}

\item{outlier}{a boolean}
}
\value{
a matrix
}
\description{
Expectation step
}
