% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{em}
\alias{em}
\title{EM Algorithm}
\usage{
em(e, m, x, K, initial_parameters, outlier = FALSE, epsilon = 1e-06)
}
\arguments{
\item{e}{a function}

\item{m}{a function}

\item{x}{a matrix}

\item{K}{a number}

\item{initial_parameters}{a list}

\item{outlier}{a boolean}

\item{epsilon}{a number}
}
\value{
a list
}
\description{
EM Algorithm
}
