% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit.mixturemodel}
\alias{fit.mixturemodel}
\title{Fit function}
\usage{
\method{fit}{mixturemodel}(object)
}
\arguments{
\item{object}{a mixturemodel object}
}
\value{
a mixturemodel object
}
\description{
Fit function
}
