% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gaussian_densities}
\alias{gaussian_densities}
\title{Gaussian densities}
\usage{
gaussian_densities(x, parameters)
}
\arguments{
\item{x}{a matrix}

\item{parameters}{a list}
}
\value{
a matrix
}
\description{
Gaussian densities
}
