% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialization}
\alias{initialization}
\title{Initialization}
\usage{
initialization(x, K, outlier = FALSE)
}
\arguments{
\item{x}{a matrix}

\item{K}{a number}

\item{outlier}{a boolean}
}
\value{
a list
}
\description{
Initialization
}
