% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{likehood}
\alias{likehood}
\title{likehood}
\usage{
likehood(e, x, parameters, outlier = FALSE)
}
\arguments{
\item{e}{a function}

\item{x}{a matrix}

\item{parameters}{a list}

\item{outlier}{a boolean}
}
\value{
a number
}
\description{
likehood
}
