% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mixte_densities}
\alias{mixte_densities}
\title{Mixte densities}
\usage{
mixte_densities(x, parameters)
}
\arguments{
\item{x}{a matrix}

\item{parameters}{a list}
}
\value{
a matrix
}
\description{
Mixte densities
}
