% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MixtureModel}
\alias{MixtureModel}
\title{Constructor function}
\usage{
MixtureModel(data, K = NULL, initialization_steps = 20)
}
\arguments{
\item{data}{a matrix}

\item{K}{a number}

\item{initialization_steps}{a number}
}
\value{
None
}
\description{
Constructor function
}
