% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{multistart_em}
\alias{multistart_em}
\title{Multistart EM}
\usage{
multistart_em(e, m, x, K, steps = 20, outlier = FALSE)
}
\arguments{
\item{e}{a function}

\item{m}{a function}

\item{x}{a matrix}

\item{K}{a number}

\item{steps}{a number}

\item{outlier}{a boolean}
}
\value{
a list
}
\description{
Multistart EM
}
