% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniform_densities}
\alias{uniform_densities}
\title{Uniform densities}
\usage{
uniform_densities(x, parameters)
}
\arguments{
\item{x}{a matrix}

\item{parameters}{a list}
}
\value{
a vector
}
\description{
Uniform densities
}
