% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{k_selection}
\alias{k_selection}
\title{Select K}
\usage{
k_selection(e, m, x, K_max)
}
\arguments{
\item{e}{a function}

\item{m}{a function}

\item{x}{a matrix}

\item{K_max}{a number}
}
\value{
a vector
}
\description{
Select K
}
