% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{m_step}
\alias{m_step}
\title{Maximization step}
\usage{
m_step(x, Q, K, outlier = FALSE)
}
\arguments{
\item{x}{a matrix}

\item{Q}{a matrix}

\item{K}{number of cluster}

\item{outlier}{a boolean}
}
\value{
a list
}
\description{
Maximization step
}
